+++ script.js
@@ -1,3 +1,6 @@
+import Book from './module/books.js';
+import Interface from './module/interface.js';
+
 const title = document.querySelector('.title');
 const author = document.querySelector('.author');
 const btn = document.querySelector('.btn');
@@ -10,51 +13,6 @@ const contact = document.querySelector('.contact');
 const navList = document.querySelector('.nav-list');
 const navAdd = document.querySelector('.nav-add');
 const navContact = document.querySelector('.nav-contact');
-const storage = JSON.parse(localStorage.getItem('form')) || [];
-
-// function Books(title, author) {
-//   this.title = title;
-//   this.author = author;
-//   this.id = Date.now();
-// }
-
-class Interface {
-  static displayBook() {
-    allBooks.innerHTML = '';
-    storage.forEach((book) => {
-      const bookAbove = document.createElement('div');
-      bookAbove.classList.add('class-book');
-      const bookAboveText = document.createElement('p');
-      bookAboveText.textContent = `"${book.title}" by ${book.author}`;
-      const bookAboveBtn = document.createElement('button');
-      bookAboveBtn.textContent = 'Remove';
-      bookAboveBtn.classList.add('remove');
-      bookAboveBtn.setAttribute('id', book.id);
-      bookAbove.append(bookAboveText);
-      bookAbove.append(bookAboveBtn);
-      allBooks.append(bookAbove);
-    });
-  }
-
-  static removeBook(targetId) {
-    const newBookList = storage.filter((book) => book.id !== targetId);
-    storage.length = 0;
-    storage.push(...newBookList);
-    localStorage.setItem('form', JSON.stringify(storage));
-    Interface.displayBook();
-  }
-
-  static appendBook(theBook) {
-    storage.push(theBook);
-    localStorage.setItem('form', JSON.stringify(storage));
-    Interface.displayBook();
-  }
-
-  static displayTime() {
-    const x = new Date();
-    document.getElementById('time').innerHTML = x;
-  }
-}
 
 document.addEventListener('DOMContentLoaded', Interface.displayBook(), Interface.displayTime());
 setInterval(Interface.displayTime, 1000);
@@ -62,7 +20,7 @@ setInterval(Interface.displayTime, 1000);
 btn.addEventListener('click', (e) => {
   e.preventDefault();
   if (title.value && author.value) {
-    const bookObj = new Books(title.value, author.value, Date.now());
+    const bookObj = new Book(title.value, author.value, Date.now());
     Interface.appendBook(bookObj);
     Interface.displayBook();
     title.value = '';
